# docs/OpenAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Storage Service",
    "description": "",
    "version": "3.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "tags": [
    {
      "name": "Categorias",
      "description": "Operações acerca das categorias de produto"
    },
    {
      "name": "Produtos",
      "description": "Operações acerca do produto"
    }
  ],
  "paths": {
    "/products": {
      "get": {
        "tags": [
          "Produtos"
        ],
        "summary": "Retorna uma lista de produtos. Utiliza paginação",
        "operationId": "buscarProdutos",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "name"
            }
          },
          {
            "name": "ascending",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Produtos"
        ],
        "summary": "Registra um novo tipo de produto",
        "operationId": "cadastrarProduto",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/products/{id_product}": {
      "get": {
        "tags": [
          "Produtos"
        ],
        "summary": "Retorna informações de um produto específico",
        "operationId": "buscarProduto",
        "parameters": [
          {
            "name": "id_product",
            "in": "path",
            "description": "Identificador do produto-alvo",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Produtos"
        ],
        "summary": "Incrementa uma quantidade de produtos no estoque",
        "operationId": "adicionarProdutos",
        "parameters": [
          {
            "name": "id_product",
            "in": "path",
            "description": "Identificador do produto-alvo",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "quantity",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 122
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Produtos"
        ],
        "summary": "Remove um produto do estoque",
        "operationId": "retirarProduto",
        "parameters": [
          {
            "name": "id_product",
            "in": "path",
            "description": "Identificador do produto-alvo",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "quantity",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 15
          },
          {
            "name": "delete",
            "in": "query",
            "description": "Se true, removerá o produto do estoque e suas informações, mas não seu historico",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Produtos"
        ],
        "summary": "Edita informações de um produto específico",
        "operationId": "atualizarProduto",
        "parameters": [
          {
            "name": "id_product",
            "in": "path",
            "description": "Identificador do produto-alvo",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/categories": {
      "get": {
        "tags": [
          "Categorias"
        ],
        "summary": "Retorna todas as categorias disponiveis",
        "operationId": "buscarCategorias",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Categorias"
        ],
        "summary": "Cria uma nova categoria de produto",
        "operationId": "criarCategoria",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/categories/{id_category}": {
      "get": {
        "tags": [
          "Categorias"
        ],
        "summary": "Busca uma categoria específica",
        "operationId": "buscarCategoria",
        "parameters": [
          {
            "name": "id_category",
            "in": "path",
            "description": "Identificador da categoria-alvo",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Categorias"
        ],
        "summary": "Exclui uma categoria",
        "operationId": "excluirCategoria",
        "parameters": [
          {
            "name": "id_category",
            "in": "path",
            "description": "Identificador da categoria-alvo",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Categorias"
        ],
        "summary": "Atualiza dados de uma categoria específica",
        "operationId": "editarCategoria",
        "parameters": [
          {
            "name": "id_category",
            "in": "path",
            "description": "Identificador da categoria-alvo",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateProductDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "categoryName": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CreateCategoryDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "UpdateProductDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "categoryName": {
            "type": "string"
          },
          "price": {
            "type": "number"
          }
        }
      },
      "UpdateCategoryDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "ProductDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CategoryDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      }
    }
  }
}